requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
django-ckeditor==6.7.3
django-cors-headers==4.9.0
django-js-asset==3.1.2
django-solo==2.5.1
djangorestframework==3.16.1
idna==3.11
pillow==12.0.0
pyTelegramBotAPI==4.29.1
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
